bunch of updates and new scrapers
- new infinite loop stats scraper
- new songs cache scraper that no longer requires direct mysql access
- new broadcaster settings for limiting requests by steps type and difficulty meter
- backend support for stepstype/difficulty song requests (coming in future update)
- normal and random requests respect current broadcaster limitations
- new !complete and !theusual commands
- other minor code updates

+++ request_list/banners.php
@@ -1,13 +1,17 @@
 <?php
 
-include ("config.php");
+include ('config.php');
+
+$fileSizeMax = 5242880; //5MB
 	
 // recieve upload of banner images via POST, FILES
-if (!isset($_POST['security_key']) || $_POST['security_key'] != $security_key){die("Fuck off");}
+if(!isset($_GET["security_key"]) || $_GET["security_key"] != $security_key || empty($_GET["security_key"])){
+    die("Fuck off");
+}
 
 $uploadfile = $uploaddir .'/'. $_FILES['file_contents']['name'];
 
-If (!file_exists($uploadfile)){
+If (!file_exists($uploadfile) && $_FILES['file_contents']['size'] <= $fileSizeMax){
 	if (move_uploaded_file($_FILES['file_contents']['tmp_name'], $uploadfile)) {
 		echo "Successfully uploaded banner for ".$_FILES['file_contents']['name']."\n";
 	}else{
